Centralize OIDC Flow to handler, and refresh logic (#17082)
* Centralize OIDC Flow to handler, and refresh logic

* Remove forced condition

* Return on success

(cherry picked from commit ebdd7f7fd9c11f1ba41e1cc4f95e222dfbab8d45)
