+++ openmetadata-spec/src/main/resources/json/schema/security/client/oidcClientConfig.json
@@ -56,6 +56,11 @@
       "type": "string",
       "enum": ["client_secret_basic", "client_secret_post", "client_secret_jwt", "private_key_jwt"]
     },
+    "tokenValidity": {
+      "description": "Validity for the JWT Token created from SAML Response",
+      "type": "integer",
+      "default": "3600"
+    },
     "customParams": {
       "description": "Custom Params.",
       "existingJavaType" : "java.util.Map<String,String>",
